requirements.txt
customtkinter>=5.2.0
CTkMessagebox>=1.6.0
deep-translator>=1.11.4
transformers>=4.41.0
Pillow>=10.3.0
speechrecognition==3.10.3
pyttsx3>=2.90
googletrans==4.0.0rc1
